ai_service/requirements.txt
fastapi
uvicorn[standard]
redis
python-multipart
pydantic
mysql-connector-python
sqlalchemy
pymysql
pytesseract
Pillow
transformers
torch
scipy
groq
